Implemented Pull to refresh on voting actvity list view.
The implementation is poor since it makes the same database queries that have been already made.
A shorter route is possible but can cause null pointer exceptions.

Also, the list view has not been tested to see if the new restaurant has actually been populated.
This can be done only when we test it across emulators.